refactor folders structure of api
