+++ frontend/src/apis/bookmark/addTagsApi.js
@@ -1,0 +1,7 @@
+import { axiosInstance } from "../api-config";
+
+export const addTagsApi = {
+  post(data) {
+    return axiosInstance.post('/api/bookmark/add', { url: data.url, tags: data.tags });
+  }
+}
