+++ frontend/src/apis/bookmark/listAllApi.js
@@ -1,0 +1,7 @@
+import { axiosInstance } from "../api-config";
+
+export const listAllBookmarkApi = {
+  post() {
+    return axiosInstance.post('/api/bookmark/list');
+  }
+}
