+++ frontend/src/apis/bookmark/listAllByTagApi.js
@@ -1,0 +1,7 @@
+import { axiosInstance } from "../api-config";
+
+export const listAllByTagApi = {
+  post(tag) {
+    return axiosInstance.post('/api/bookmark/list_by_tag', { tag: tag });
+  }
+}
