+++ frontend/src/apis/bookmark/newBookmarkApi.js
@@ -1,0 +1,7 @@
+import { axiosInstance } from "../api-config";
+
+export const newBookmarkApi = {
+  post(url) {
+    return axiosInstance.post('/api/bookmark/add', { url: url });
+  }
+}
