+++ frontend/src/apis/bookmark/recentApi.js
@@ -1,0 +1,7 @@
+import { axiosInstance } from "../api-config";
+
+export const recentViewedApi = {
+  post() {
+    return axiosInstance.post('/api/bookmark/recent');
+  }
+}
