+++ frontend/src/apis/bookmark/removeBookmark.js
@@ -1,0 +1,8 @@
+import { axiosInstance } from './../api-config';
+
+export const removeBookmarkApi = {
+  post(url) {
+    return axiosInstance.post('/api/bookmark/remove', { url: url });
+
+  }
+}
