+++ frontend/src/apis/index.js
@@ -1,0 +1,4 @@
+export * from "./api-config"
+export * as bookmarkApi from "./bookmark"
+export * as snapshotApi from "./snapshot"
+export * as userApi from "./user"
