+++ frontend/src/apis/snapshot/listSnapshotVersionsApi.js
@@ -1,4 +1,4 @@
-import { axiosInstance } from "./api-config";
+import { axiosInstance } from "../api-config";
 
 export const listSnapshotVersionsApi = {
   get(params) {
