+++ frontend/src/apis/user/loginApi.js
@@ -1,4 +1,4 @@
-import { axiosInstance } from "./api-config";
+import { axiosInstance } from "../api-config";
 
 export const loginApi = {
   post(data) {
